Temporarily include llvm 14 libraries in order to switch to llvm 15
This is a hack to make it possible to switch EPEL package builds over
from llvm 14 to 15.

RHEL doesn't ship a compat version of llvm 14 and we need to make both
versions available at the same time somehow in order to build ldc with
itself and switch over to llvm 15 that's the default in RHEL 9.2.

The canonical way to do it would be to add a llvm14 compat package, but
in this case it's only needed to be able to perform a package rebuild in
koji (we don't need to make it available to end users) and it's not
worth adding the compat package only for that.